Initial commit of aws_rds_instance resource
This is a bare-bones resource for creating and destroying RDS instances.
RDS instances have a lot of possible options, so rather than list them
all in the resources, I've provided an `additional_options` hash that
the user can use to directly pass options to the underlying driver.  We
will likely expand the functionality here as we use it ourselves.