Fix handling of lb server certificates
Previously, if a user specified an ARN for the `server_certificate` key
in a listener spec, the LB creation would succeed, but future updates
would always update the server_certificate.  If the user specified an
ARN for the `ssl_certificate_id`, the creation would succeed, but future
updates would fail.  Here we support both.